requirements.txt
Django>=5.0
djangorestframework
gunicorn
firebase-admin
whitenoise

